refactor: split symbiont project file helpers
